Mass Name Change
By changing the name from Bitcoin to Genex, this would make it the first step to defining Genex from Bitcoin. If you was able to run this though, it would still run of Bitcoins preexisting chain.
There is still yet a lot of work to be done before it runs off its own Blockchain.
3 name changes took place, all "match case":
-Bitcoin = Genex
-bitcoin = genex
-BITCOIN = GENEX