Changed Daemon Port
Bitcoins Daemon port used to run on :8333. Now Genexs Daemon port runs on :17070
We set it as an odd number because different programs use different ports. If 2 ports was the same, you wouldnt be able to make it work as the other program is using the port already. So by setting it as a number 10,000> its hardly like another program is going to be using that port as main programs tend to use 1000<