New Algorithm
Implentation of Lyra2RE algorithm, will change to v2 when figured out how.
Now this algorithm is needed because we have ASICs in the original Bitcoin algorithm, SHA256, so to prevent this we have opt'd for an algorithm that doesn't take to kindly to them when they attempt to mine it.
Hopefully by these changes the algorithm change should be a success